Rename TheGame2 to GameScreen and modify to fit
